requirements.txt
pandas
pyyaml
requests
pyarrow
quarto
